% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_psag.R
\name{fast_psag}
\alias{fast_psag}
\title{fast PSA-G}
\usage{
fast_psag(X, n_components = NULL, percent_appx = 0.99)
}
\arguments{
\item{X}{data matrix. May have column names.}

\item{n_components}{dimension of compositions to which the data is reduced by PCA}

\item{percent_appx}{percent of variance to be explained by PCA}
}
\value{
A list of the following components of PSA-G.
\item{X}{the input matrix.}
\item{Xhat}{a list of lower dimensional representations with respect to the original basis.}
\item{Vhat}{a list of matrix representing vertices of the lower dimensional subsimplex.}
\item{Xhat_reduced}{a list of lower dimensional representations with respect to the reduced basis \code{Vhat}}
\item{scores}{a matrix of scores.}
\item{residuals}{a list of residuals.}
\item{scores}{a matrix of scores.}
\item{RSS}{a vector of residual sums of squares.}
\item{backwards_mean}{the backwards mean. Equal to \code{Vhat$'r=0'}.}
\item{modes}{a matrix of loading vectors.}
\item{construction_info}{a list of \code{idx} and \code{a}. \code{idx} contains the index of
the vertex that were chosen at each step. \code{a} contains the corresponding normal vector.}
}
\description{
fast PSA-G
}
