% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loading.R
\name{plot_bar}
\alias{plot_bar}
\title{Single Bar Plot for PSA Loadings or Vertices}
\usage{
plot_bar(v, max.k = 12)
}
\arguments{
\item{v}{a loading vector}

\item{max.k}{maximum number of elements to display}
}
\value{
a ggplot object
}
\description{
Creating a bar plot summarizing a loading vector or a vertex vector
}
