% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loading.R
\name{plot_loading}
\alias{plot_loading}
\title{Loading Plot for PSA}
\usage{
plot_loading(psa.res, k = 4, max.k = 12)
}
\arguments{
\item{psa.res}{output of \code{psa()}}

\item{k}{number of loading vectors to display}

\item{max.k}{maximum number of elements to display}
}
\value{
a ggplot object
}
\description{
Creating a bar plot summarizing a loading vectors of PSA
}
