% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ternary.R
\name{plot_vertex}
\alias{plot_vertex}
\title{Vertices Plot for PSA}
\usage{
plot_vertex(V, max.k = 12)
}
\arguments{
\item{V}{a matrix of vertices (each row representing a vertex)}

\item{max.k}{maximum number of elements to display}
}
\value{
a ggplot object
}
\description{
Creating bar plots representing vertices of PSA
}
