% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa.R
\name{psa}
\alias{psa}
\title{Principal Nested Simplices Analysis}
\usage{
psa(type, X, testweights = seq(0, 1, length.out = 100))
}
\arguments{
\item{type}{\code{s} for PSA-S or \code{o} for PSA-O.}

\item{X}{A data matrix.}

\item{testweights}{A vector of weights to try.}
}
\value{
A list
\itemize{
\item vertices a list of matrix representing vertices of the lower dimensional subsimplex.
The $r$th element of the list corresponds to the rank $r-1$ subsimplex.
\item pts a list of lower dimensional representations with respect to the reduced basis \code{vertices}
\item pts.approx a list of lower dimensional representations with respect to the original basis
\item scores a matrix of scores.
\item rss a vector of residual sums of squares.
\item modes a list of modes of variation. The $r$th element of the list is the difference of
rank $r$ approximations to rank $r-1$ approximations.
\item loadings a matrix of loading vectors.
\item const.info a data frame of merged vertices and merging weight at each merge.
\item dendrogram.input .
}
}
\description{
Estimate PSA-S or PSA-O of given data matrix.
}
