% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psacomp-package.R
\docType{package}
\name{psacomp-package}
\alias{psacomp}
\alias{psacomp-package}
\title{psacomp: Principal Subsimplex Analysis for Compositional Data}
\description{
"psacomp'' is a package for computing Principal Subsimplex Analysis introduced in Principal Subsimplex Analysis (2025+), Hyeon Lee, Kassel Liam Hingee, Janice L. Scealy, Andrew T. A. Wood, Eric Grunsky, and J. S. Marron.
}
\author{
\strong{Maintainer}: Hyeon Lee \email{haneone19997@gmail.edu}

Authors:
\itemize{
  \item Kassel Liam Hingee
}

Other contributors:
\itemize{
  \item Janice Scealy [contributor]
  \item Andrew Wood [contributor]
  \item Eric Grunsky [contributor]
  \item J. S. Marron [contributor]
}

}
\keyword{internal}
