% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psag.R
\name{psag}
\alias{psag}
\title{PSA-G}
\usage{
psag(X)
}
\arguments{
\item{X}{data matrix. May have column names.}
}
\value{
A list of the following components of PSA-G.
\item{X}{the input matrix.}
\item{Xhat}{a list of lower dimensional representations with respect to the original basis.}
\item{Vhat}{a list of matrix representing vertices of the lower dimensional subsimplex.}
\item{Xhat_reduced}{a list of lower dimensional representations with respect to the reduced basis \code{Vhat}}
\item{scores}{a matrix of scores.}
\item{residuals}{a list of residuals.}
\item{scores}{a matrix of scores.}
\item{RSS}{a vector of residual sums of squares.}
\item{backwards_mean}{the backwards mean. Equal to \code{Vhat$'r=0'}.}
\item{modes}{a matrix of loading vectors.}
\item{construction_info}{a list of \code{idx} and \code{a}. \code{idx} contains the index of
the vertex that were chosen at each step. \code{a} contains the corresponding normal vector.}
}
\description{
PSA-G
}
