% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ternary.R
\name{plot_ternary}
\alias{plot_ternary}
\title{Ternary Plot for PSA}
\usage{
plot_ternary(psa.res, group = NULL, max.k = 12)
}
\arguments{
\item{psa.res}{output of \code{psa()}}

\item{group}{a vector of factors representing groups of data points if applicable}

\item{max.k}{maximum number of elements to display}
}
\value{
a ggplot object
}
\description{
Creating a ternary plot with its vertices for 2-dimensional
representation of obtained from PSA
}
