% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psa.R
\name{psa}
\alias{psa}
\title{Principal Subsimplex Analysis}
\usage{
psa(type, X, testweights = seq(0, 1, length.out = 100))
}
\arguments{
\item{type}{'s' for PSA-S or 'o' for PSA-O.}

\item{X}{a data matrix.}

\item{testweights}{a vector of weights for grid search for alpha.}
}
\value{
A list of the following components of PSA.
\item{Vhat}{a list of matrix representing vertices of the lower dimensional subsimplex.}
\item{Xhat}{a list of lower dimensional representations with respect to the original basis.}
\item{Xhat_reduced}{a list of lower dimensional representations with respect to the reduced basis \code{Vhat}}
\item{scores}{a matrix of scores.}
\item{X}{the input matrix.}
\item{residuals}{a list of residuals.}
\item{scores}{a matrix of scores.}
\item{RSS}{a vector of residual sums of squares.}
\item{backwards_mean}{the backwards mean. Equal to \code{Vhat$'r=0'}.}
\item{loadings}{a matrix of loading vectors.}
\item{construction_info}{a data frame of merged vertices and merging weight at each merge.}
}
\description{
Estimate PSA-S or PSA-O of given data matrix.
}
